+++ index.js
@@ -1,16 +1,16 @@
 'use strict';
 
-var bluebird= require('bluebird');
+// ensure process is running on linux
+if(process.platform !== 'linux') {
+	throw new Error('linux-user must be running on Linux');
+}
+
+// ensure process is running as root
+if(!(process.getuid && process.getuid() === 0)) {
+	throw new Error('linux-user must be running as root user');
+}
+
 var lib = require('./lib/user');
-
-lib.promise = (function(lib) {
-	var out = {};
-
-	Object.keys(lib).forEach(function(key){
- 		out[key] = bluebird.promisify(lib[key]);
-	});
-
-	return out;
-})(lib);
+lib.promises = require('./lib/promise')
 
 module.exports = lib;
