+++ lib/user.js
@@ -5,22 +5,12 @@ var exec = require('child_process').exec;
 var spawn = require('child_process').spawn;
 var fs = require('fs');
 
-if(process.platform !== 'linux') {
-  throw new Error('linux-user must be running on Linux');
-}
-
-// ensure process is running as root
-if(!(process.getuid && process.getuid() === 0)) {
-  throw new Error('linux-user must be running as root user');
-}
-
 var validUsernameRegex = /^([a-z_][a-z0-9_-]{0,30})$/;
 
 function validate (username) {
-  return validUsernameRegex.test(username);
+     return validUsernameRegex.test(username);
 }
 
-exports.validateUsername = validate;
 
 var user_arg_map = {
   shell: function(SHELL){
@@ -296,3 +286,6 @@ exports.addSSHtoUser = function (user, key, callback){
     });
   });
 };
+
+
+exports.validateUsername = validate;
