+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "linux-sys-user",
-  "version": "0.6.3",
+  "version": "1.0.0",
   "description": "Node module for Linux user and group control.",
   "author": [
     {
@@ -43,7 +43,5 @@
     "type": "git",
     "url": "https://github.com/wmantly/linux-user.git"
   },
-  "dependencies": {
-    "bluebird": "^3.5.5"
-  }
+  "dependencies": {}
 }
